+++ matjalalzz/src/main/java/shop/matjalalzz/chat/config/WebSocketConfig.java
@@ -7,6 +7,9 @@ import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
+import shop.matjalalzz.chat.interceptor.HandshakeAuthInterceptor;
+import shop.matjalalzz.chat.interceptor.WebSocketAuthInterceptor;
+import shop.matjalalzz.chat.interceptor.WebSocketLoggingInterceptor;
 
 @Configuration
 @EnableWebSocketMessageBroker
