+++ matjalalzz/src/main/java/shop/matjalalzz/chat/interceptor/HandshakeAuthInterceptor.java
@@ -1,4 +1,4 @@
-package shop.matjalalzz.chat.config;
+package shop.matjalalzz.chat.interceptor;
 
 import java.net.URI;
 import java.util.Map;
